safety-monitor/requirements.txt
Flask==3.0.3
Flask_SocketIO==5.3.6
meteocalc==1.1.0
pip==24.0
psutil==5.9.8
python-dotenv==1.0.1
pytz==2024.1
Requests==2.31.0
zeroconf==0.132.2
